Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240821/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Main {
 
     public static long M;
-    public static int a;
+    public static long a;
     public static long b;
 
     public static int minCount = 19;
@@ -41,7 +41,7 @@ public class Main {
 
         M = sc.nextLong();
 
-        a = sc.nextInt();
+        a = sc.nextLong();
         b = sc.nextLong();
 
         for (int i = a; i <= b; i++) {
